把接口添加进权限控制

+++ app/Http/Controllers/API/VideoController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers\API;
 use App\Models\Video;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-
+use Illuminate\Support\Facades\Auth;
 
 class VideoController extends Controller
 {
+
+
     public function uploadVideo () {
+
         /*
         接收文件并合并
         */
@@ -30,6 +33,12 @@ class VideoController extends Controller
 
     public function insert(Request $request){
 
+        $user  = Auth::user();
+
+        if (!($user->hasRole('teacher') || $user->hasRole("admin"))) {
+            return response()->json(['message' => 'unauthorized '],401);
+        }
+
         $userid = $request->input('userid');
         $cid = $request->input('cid');
         $filename = $request->input('filename');
@@ -45,18 +54,44 @@ class VideoController extends Controller
     }
 
     public function selectForUserid(Request $request) {
+
+        $user  = Auth::user();
+
+        if (!($user->hasRole('teacher') || $user->hasRole("admin"))) {
+            return response()->json(['message' => 'unauthorized '],401);
+        }
+
+        $user = Auth::user();
+        dd($user);
+        $flag = $user->hasRole('admin');
+        dd($flag);
         $userid = $request->input('userid');
         $video = new Video();
         $data = $video->where("userid",$userid)->get()->toJson();
         return $data;
     }
     public function selectForCid(Request $request) {
+
+        $user  = Auth::user();
+
+        if (!($user->hasRole('teacher') || $user->hasRole("admin"))) {
+            return response()->json(['message' => 'unauthorized '],401);
+        }
+
         $cid = $request->input('cid');
         $video = new Video();
         $data = $video->where("cid",$cid)->get()->toJson();
         return $data;
     }
     public function delete(Request $request) {
+
+        $user  = Auth::user();
+
+        if (!($user->hasRole('teacher') || $user->hasRole("admin"))) {
+            return response()->json(['message' => 'unauthorized '],401);
+        }
+
+
         $id = $request->input('id');
         $video = Video::find($id);
         if ($video->delete()){
@@ -71,6 +106,13 @@ class VideoController extends Controller
     }
 
     public function update(Request $request) {
+
+        $user  = Auth::user();
+
+        if (!($user->hasRole('teacher') || $user->hasRole("admin"))) {
+            return response()->json(['message' => 'unauthorized '],401);
+        }
+
         $id = $request->input('id');
         $videoName = $request->input('videoName');
         $kp = $request->input('kp');
@@ -81,4 +123,6 @@ class VideoController extends Controller
             return json_encode(["status"=>0,"message"=>"删除成功！"]);
         }
     }
+
+
 }
